refactor: move mlu linalg utilities to util.
