+++ xllm/core/layers/mlu/CMakeLists.txt
@@ -9,12 +9,11 @@ cc_library(
     deepseek_v2_decoder_layer_impl.h
     deepseek_v2_sparse_moe_block.h
     indexer.h
-    linalg.h
     deepseek_v4/compressor.h
     deepseek_v4/dsa_cache_mapping.h
     deepseek_v4/dsa_metadata_builder_mlu.h
-    deepseek_v4/indexer.h
-    deepseek_v4/attention.h
+    deepseek_v4/deepseek_v4_indexer.h
+    deepseek_v4/deepseek_v4_attention.h
     deepseek_v32_sp_context.h
     deepseek_v32_sp_plan.h
     moe_gate.h
@@ -29,11 +28,10 @@ cc_library(
     deepseek_v2_decoder_layer_impl.cpp
     deepseek_v2_sparse_moe_block.cpp
     indexer.cpp
-    linalg.cpp
     deepseek_v4/compressor.cpp
     deepseek_v4/dsa_metadata_builder_mlu.cpp
-    deepseek_v4/indexer.cpp
-    deepseek_v4/attention.cpp
+    deepseek_v4/deepseek_v4_indexer.cpp
+    deepseek_v4/deepseek_v4_attention.cpp
     moe_gate.cpp
     fused_moe.cpp
     fused_moe_base.cpp
@@ -43,4 +41,5 @@ cc_library(
     qwen3_5_decoder_layer.cpp
   DEPS
     :common_layers
+    :util
 )
